chore: move types to types file, other random changes
