+++ maiar-client/src/components/GridLayout.tsx
@@ -28,21 +28,21 @@ type LayoutType = {
 // Define the layout for different breakpoints
 const DEFAULT_LAYOUTS: LayoutType = {
   lg: [
-    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 12, minH: 3 },
+    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 3, minH: 3 },
     { i: "pipeline", x: 0, y: 2, w: 4, h: 12, minW: 3, minH: 6 },
     { i: "contextChain", x: 4, y: 2, w: 4, h: 12, minW: 3, minH: 6 },
     { i: "chat", x: 8, y: 2, w: 4, h: 6, minW: 3, minH: 4 },
     { i: "events", x: 8, y: 8, w: 4, h: 6, minW: 3, minH: 4 }
   ],
   md: [
-    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 12, minH: 3 },
+    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 3, minH: 3 },
     { i: "pipeline", x: 0, y: 2, w: 6, h: 12, minW: 3, minH: 6 },
     { i: "contextChain", x: 6, y: 2, w: 6, h: 12, minW: 3, minH: 6 },
     { i: "chat", x: 0, y: 14, w: 6, h: 6, minW: 3, minH: 4 },
     { i: "events", x: 6, y: 14, w: 6, h: 6, minW: 3, minH: 4 }
   ],
   sm: [
-    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 12, minH: 3 },
+    { i: "status", x: 0, y: 0, w: 12, h: 4, minW: 3, minH: 3 },
     { i: "pipeline", x: 0, y: 2, w: 12, h: 8, minW: 3, minH: 6 },
     { i: "contextChain", x: 0, y: 10, w: 12, h: 8, minW: 3, minH: 6 },
     { i: "chat", x: 0, y: 18, w: 12, h: 6, minW: 3, minH: 4 },
