+++ packages/model-openai/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./openai";
+export * from "./types";
