+++ packages/plugin-image/src/plugin.ts
@@ -1,23 +1,6 @@
 import { PluginBase, AgentContext, PluginResult } from "@maiar-ai/core";
 import { PromptResponseSchema } from "./types";
 import { generatePromptTemplate } from "./templates";
-import { z } from "zod";
-
-export const imageGenerationSchema = {
-  input: z.string(),
-  output: z.array(z.string())
-};
-
-type ImageGenerationCapability = {
-  input: string;
-  output: string[];
-};
-
-declare module "@maiar-ai/core" {
-  export interface ICapabilities {
-    "image-generation": ImageGenerationCapability;
-  }
-}
 
 export class PluginImageGeneration extends PluginBase {
   constructor() {
